# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WavesOfEgypt travel marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tours:
    get:
      operationId: listTours
      tags: [tours]
      summary: List tours with optional filters
      parameters:
        - name: destinationId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating, newest, best_seller, duration_asc, duration_desc]
        - name: isPrivate
          in: query
          schema:
            type: boolean
        - name: isFamilyFriendly
          in: query
          schema:
            type: boolean
        - name: hasHotelPickup
          in: query
          schema:
            type: boolean
        - name: freeCancellation
          in: query
          schema:
            type: boolean
        - name: instantConfirmation
          in: query
          schema:
            type: boolean
        - name: minDuration
          in: query
          schema:
            type: number
        - name: maxDuration
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: isBestSeller
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of tours
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourList"

    post:
      operationId: createTour
      tags: [tours]
      summary: Create a tour (vendor/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"
  /tours/{id}:
    get:
      operationId: getTour
      tags: [tours]
      summary: Get tour by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tour details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    patch:
      operationId: updateTour
      tags: [tours]
      summary: Update a tour (vendor/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourUpdate"
      responses:
        "200":
          description: Updated tour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tour"

    delete:
      operationId: deleteTour
      tags: [tours]
      summary: Delete a tour (vendor/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tours/featured:
    get:
      operationId: getFeaturedTours
      tags: [tours]
      summary: Get featured tours for homepage
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Featured tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
  /tours/best-sellers:
    get:
      operationId: getBestSellerTours
      tags: [tours]
      summary: Get best-selling tours
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Best sellers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"
...
  /tours/{id}/availability:
    get:
      operationId: getTourAvailability
      tags: [tours]
      summary: Get tour availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlot"

  # ── REVIEWS ──────────────────────────────────────────────────────────────
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get current user's wishlist
      responses:
        "200":
          description: Wishlist tours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tour"

    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add a tour to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /wishlist/{tourId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove tour from wishlist
      parameters:
        - name: tourId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── STATS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── AUTH ────────────────────────────────────────────────────────────────
...
    Tour:
      type: object
      required: [id, title, slug, price, durationHours, destinationId, categoryId, rating, reviewCount, coverImage, isAvailable]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        durationHours:
          type: number
        destinationId:
          type: integer
        destinationName:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        included:
          type: array
          items:
            type: string
        excluded:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        maxParticipants:
          type: ["integer", "null"]
        minAge:
          type: ["integer", "null"]
        difficulty:
          type: ["string", "null"]
          enum: [easy, moderate, challenging, null]
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        isPrivate:
          type: boolean
        isFamilyFriendly:
          type: boolean
        hasHotelPickup:
          type: boolean
        freeCancellation:
          type: boolean
        instantConfirmation:
          type: boolean
        groupType:
          type: string
          enum: [private, group]
        badge:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    TourDetail:
      allOf:
        - $ref: "#/components/schemas/Tour"
        - type: object
          properties:
            itinerary:
              type: array
              items:
                $ref: "#/components/schemas/ItineraryStep"
            meetingPoint:
              type: ["string", "null"]
            languages:
              type: array
              items:
                type: string
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            faq:
              type: array
              items:
                type: object
                properties:
                  question:
                    type: string
                  answer:
                    type: string
            whatsappNumber:
              type: ["string", "null"]
            videoUrl:
              type: ["string", "null"]
    TourList:
      type: object
      required: [tours, total, limit, offset]
      properties:
        tours:
          type: array
          items:
            $ref: "#/components/schemas/Tour"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    TourInput:
      type: object
      required: [title, slug, price, durationHours, destinationId, categoryId, coverImage]
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        durationHours:
          type: number
        destinationId:
          type: integer
        categoryId:
          type: integer
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        included:
          type: array
          items:
            type: string
        excluded:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        maxParticipants:
          type: ["integer", "null"]
        minAge:
          type: ["integer", "null"]
        difficulty:
          type: string
          enum: [easy, moderate, challenging]
        meetingPoint:
          type: ["string", "null"]
        languages:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
    TourUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        durationHours:
          type: number
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        included:
          type: array
          items:
            type: string
        excluded:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        maxParticipants:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
    ItineraryStep:
      type: object
      required: [order, title]
      properties:
        order:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
    AvailabilitySlot:
      type: object
      required: [date, available, spotsLeft]
      properties:
        date:
          type: string
          format: date
        available:
          type: boolean
        spotsLeft:
          type: integer
        price:
          type: ["number", "null"]

    # ── REVIEW ───────────────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, tourId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        tourId:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        userAvatar:
          type: ["string", "null"]
        rating:
          type: number
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    WishlistInput:
      type: object
      required: [tourId]
      properties:
        tourId:
          type: integer

    # ── STATS ─────────────────────────────────────────────────────────────────
...