# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WavesOfEgypt travel marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilitySlot {
  date: Date;
  available: boolean;
  spotsLeft: number;
  /** @nullable */
  price?: number | null;
}