# lib/api-zod/src/generated/types/tourDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WavesOfEgypt travel marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryStep } from './itineraryStep';
import type { Review } from './review';
import type { Tour } from './tour';
import type { TourDetailFaqItem } from './tourDetailFaqItem';

export type TourDetail = Tour & ({
  itinerary?: ItineraryStep[];
  /** @nullable */
  meetingPoint?: string | null;
  languages?: string[];
  reviews?: Review[];
  faq?: TourDetailFaqItem[];
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  videoUrl?: string | null;
});